+++ lib/src/models/story_item.dart
@@ -59,26 +59,3 @@ class StoryItem {
   /// Applicable when [storyItemType] is [StoryItemType.web]
   final StoryViewWebConfig? webConfig;
 }
-
-class StoryCustomWidgetWrapper extends StatelessWidget {
-  const StoryCustomWidgetWrapper({
-    super.key,
-    required this.child,
-    this.isAutoStart = true,
-    this.onLoaded,
-  });
-
-  final Widget? child;
-  final bool isAutoStart;
-  final Function()? onLoaded;
-
-  @override
-  Widget build(BuildContext context) {
-    return Builder(
-      builder: (context) {
-        onLoaded?.call();
-        return child ?? const SizedBox.shrink();
-      },
-    );
-  }
-}
